Prefix VirtualHosts conf file
In an attempt to ensure the preconfigured VirtualHosts file will
override any existing files, this commit adds a numerical prefix (01_) to the
file. Admittedly, this seems a bit hamfisted and more testing would be
needed to determine the efficacy of this measure under various
circumstances, but for now, this is the noodle that stuck.